♻️ 모달 공통화 - 선택 버튼 로직 및 UI BaseModal로 이동
